fixed major bug
forgot to check if the letters is empty when attempting to update the first position of a letter. (not sure why it was not causing a crash beforehand, only when adding an unused variable int/string in MultiText caused it to crash. im assuming because this changed the memory and it finally caused a crash idk)